% tinyrox says don't edit this manually, but it can't stop you!
\docType{package}
\name{mx.client-package}
\alias{mx.client}
\alias{mx.client-package}
\title{\packageTitle{mx.client}}
\description{\packageDescription{mx.client}}
\section{Package Content}{\packageIndices{mx.client}}
\author{\packageAuthor{mx.client}}
\section{Maintainer}{\packageMaintainer{mx.client}}
