% tinyrox says don't edit this manually, but it can't stop you!
\name{mx_client_from_config}
\alias{mx_client_from_config}
\title{Wrap a list as an mx.client config}
\usage{
mx_client_from_config(cfg, path = NULL, app = NULL)
}
\arguments{
\item{cfg}{Named list.}

\item{path}{Character or NULL. Source/sink path for saves.}

\item{app}{Character or NULL. Application namespace.}
}
\value{
An object of class \code{"mx_client_config"}.
}
\description{
Wrap a list as an mx.client config
}
